新版消费机代码

+++ kill_pid.py
@@ -58,10 +58,15 @@ class ScheduleJob:
         # 判断是否是僵尸进程
         try:
 
-            return  (
-    self.machine_process.status() == psutil.STATUS_ZOMBIE
-    and self.machine_process_old.status() == psutil.STATUS_ZOMBIE
-)
+            return self.machine_process.status() == psutil.STATUS_ZOMBIE
+        except psutil.NoSuchProcess:
+            return False
+
+    def is_zombie_process_old(self):
+        # 判断是否是僵尸进程
+        try:
+
+            return self.machine_process_old.status() == psutil.STATUS_ZOMBIE
         except psutil.NoSuchProcess:
             return False
 
@@ -73,6 +78,17 @@ class ScheduleJob:
             return False
         try:
             self.machine_process.terminate()  # 或者使用 process.kill() 来强制终止进程
+            return True
+        except psutil.NoSuchProcess:
+            return False
+
+    def kill_zombie_process_old(self):
+        """
+             杀死僵尸进程
+             """
+        if not self.is_zombie_process_old():
+            return False
+        try:
             self.machine_process_old.terminate()  # 或者使用 process.kill() 来强制终止进程
             return True
         except psutil.NoSuchProcess:
@@ -83,14 +99,25 @@ class ScheduleJob:
         启动消费机工作进程
         """
         print("#" * 30)
-        p1 = multiprocessing.Process(target=MachinePublic.main, args=())
-        p2 = multiprocessing.Process(target=MachinePublic_old.main, args=())
-        p1.start()
-        p2.start()
+
+        try:
+            psutil.Process(self.machine_pid)
+        except Exception:
+            p1 = multiprocessing.Process(target=MachinePublic_old.main, args=())
+            p1.start()
+            self.machine_pid = p1.pid
+            print("I'm 正在重启进程...", self.machine_process, 'pid:', self.machine_pid)
+        try:
+            psutil.Process(self.machine_pid_old)
+        except Exception:
+            p2 = multiprocessing.Process(target=MachinePublic_old.main, args=())
+            p2.start()
+            self.machine_pid_old = p2.pid
+            print("I'm 正在重启进程...", self.machine_process_old, 'pid_old:', self.machine_pid_old)
+
+
+
         # 分别写入两个 PID 在各自脚本主方法里
-        self.machine_pid = p1.pid
-        self.machine_pid_old = p2.pid
-        print("I'm 正在重启进程...", self.machine_process, 'pid:', self.machine_pid, 'pid_old:', self.machine_pid_old)
 
 
     def check_machine_alive(self):
@@ -101,24 +128,34 @@ class ScheduleJob:
         但还是会做错误检查，如果没有错误，说明进程存在，反之进程不存在
         :return:
         """
+        if not os.path.isfile("pid.txt"):
+            with open("pid.txt", "w") as f:
+                f.write("0")
+
+        if not os.path.isfile("pid.txt"):
+            with open("pid_old.txt", "w") as f:
+                f.write("0")
         with open("pid.txt", "r") as f:
             self.machine_pid = int(f.read())
 
-        with open("pid_old.txt", "r") as z:
-            self.machine_pid_old = int(z.read())
+        with open("pid_old.txt", "r") as f:
+            self.machine_pid_old = int(f.read())
         try:
-
             self.machine_process = psutil.Process(self.machine_pid)
             self.machine_process_old = psutil.Process(self.machine_pid_old)
-            if self.kill_zombie_process():
+            if self.kill_zombie_process() or self.kill_zombie_process_old():
                 # WeChatPush(server='已成僵尸进程, 准备重启').run()
                 # 是僵尸进程就杀死当前僵尸进程，然后重新启动
                 self.start_machine_process()
 
         except psutil.NoSuchProcess:
+            self.machine_process = None
+            self.machine_process_old = None
             # WeChatPush(server='进程意外退出').run()
             self.start_machine_process()
         except Exception as e:
+            self.machine_process = None
+            self.machine_process_old = None
             traceback.print_exc()
             print("check_pid_alive函数报错", e)
             self.start_machine_process()
@@ -130,12 +167,13 @@ class ScheduleJob:
 def main():
     print("开始运行父进程", os.getpid())
     SJ = ScheduleJob()
-    schedule.every(3).seconds.do(SJ.check_machine_alive)  # 每隔3秒检查进程是否存活
+    SJ.check_machine_alive()
+    schedule.every(360).seconds.do(SJ.check_machine_alive)  # 每隔3秒检查进程是否存活
     # 每天的特定时间执行任务
     # schedule.every(1).day.at("08:00").do(WeChatPush(server='每日自检').check)
     while True:
         schedule.run_pending()  # run_pending：运行所有可以运行的任务
-        time.sleep(1)
+        time.sleep(3)
 
 
 if __name__ == '__main__':
